Ready for Demo UI work - Added User History Tab - Added Profile Placeholder - Fixed Search Function and UI - Upload Resource Fine Tuning pending - Find Treatment UI fixes
