+++ client/src/App.js
@@ -1,14 +1,14 @@
 import { useEffect, useState } from 'react'
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import { Error, Landing, Register, ProtectedRoute } from './pages'
-import { AddJob, AllJobs, Profile, Stats, SharedLayout, Upload, ArticleUpdate, EditArticle, FetchArticles, ArticleDetails } from './pages/dashboard'
+import { AddJob, AllJobs, Profile, SharedLayout, Upload, EditArticle, ArticleDetails, UserProfile } from './pages/dashboard'
+// import { Stats, ArticleUpdate, FetchArticles } from './pages/dashboard'
 import { LoadScript } from '@react-google-maps/api';
 
 
 
 function App() {
   const [data, setData] = useState([]);
-  const libraries = ['places'];
   useEffect(() => {
     fetch('/api/data')
       .then(response => response.json())
@@ -39,7 +39,7 @@ function App() {
             <Route path='getcentre' element={<EditArticle />} />
             <Route path='upload' element={<Upload />} />
             <Route path="details/:itemTitle" element={<ArticleDetails data={data} />} />
-
+            <Route path='user-profile' element={<UserProfile />} />
 
           </Route>
           <Route path='/register' element={<Register />} />
