+++ client/src/pages/dashboard/EditArticle.jsx
@@ -19,12 +19,14 @@ function EditArticle() {
 
   const [recentDocumentLoading, setRecentDocumentLoading] = useState(true);
 
+  // eslint-disable-next-line no-unused-vars
   const [selectedType, setSelectedType] = useState('search');
+  // eslint-disable-next-line no-unused-vars
   const [updateMessage, setUpdateMessage] = useState('');
   const [recentDocument, setRecentDocument] = useState(null);
   const [selectedFilterTopic, setSelectedFilterTopic] = useState('');
   const [isBrowseKnowledgeArticleClicked, setisBrowseKnowledgeArticleClicked] = useState(false);
-
+  const [searchString, setSearchString] = useState('');
   const POLLING_INTERVAL = 5000;
 
   // Function to handle resource topic change
@@ -32,9 +34,11 @@ function EditArticle() {
     const selectedFilterTopic = event.target.value;
     setSelectedFilterTopic(selectedFilterTopic);
   };
+
   const handleUpdate = async () => {
     try {
       setRecentDocumentLoading(true);
+      setSearchString(context)
       const response = await fetch('/api/update', {
         method: 'POST',
         headers: {
@@ -45,15 +49,10 @@ function EditArticle() {
       const data = await response.json();
       setUpdateMessage(data.message);
       setSearchPerformed(true);
-
-
     } catch (error) {
       console.error('Error updating article:', error);
       setUpdateMessage('An error occurred while updating the article');
     }
-
-
-
   };
 
   const toggleArticleMode = () => {
@@ -67,6 +66,10 @@ function EditArticle() {
     }
   }
 
+  // function compareByRank(a, b) {
+  //   return a.rank - b.rank;
+  // }
+
   useEffect(() => {
     let pollTimer;
 
@@ -79,9 +82,6 @@ function EditArticle() {
         const matchingTopics = Object.keys(newData).filter(topic =>
           topic.toLowerCase().includes(context.toLowerCase())
         );
-
-
-
 
         if (matchingTopics.length > 0) {
           setRecentDocument(newData);
@@ -118,8 +118,6 @@ function EditArticle() {
   }, []);
 
   const ArticleCard = ({ item }) => {
-    //  const [expanded, setExpanded] = useState(false);
-    // console.log(item)
     const imageUrl = item.thumbnails.startsWith('https://')
       ? item.thumbnails
       : `https://${item.thumbnails}`;
@@ -130,42 +128,32 @@ function EditArticle() {
         <Card className='custom-card--container'>
           <Card.Img variant="top" src={imageUrl} height={200} className='border-bottom--075 min-height--50' />
           <Card.Body>
-            <Card.Title>{item.title}</Card.Title>
+            <Card.Title>{item.title}
+              {item.url && <a target='_blank' rel="noreferrer" href={item.url}><i class="bi bi-box-arrow-up-right ms-2"></i></a>}
+            </Card.Title>
             <Card.Subtitle>{item.topic}</Card.Subtitle>
             <Card.Text className='custom-styled-card'>
               {item.summary}
             </Card.Text>
-            <Button variant="primary"><Link to={`/details/${item.title}`} onClick={()=> localStorage.setItem('article-title', item.title)}>Know More</Link></Button>
+            <Button variant="primary"><Link to={`/details/${item.title}`} onClick={() => localStorage.setItem('article-title', item.title)}>Know More</Link></Button>
+            
+            {item.url && <a className='custom-link' target='_blank' rel="noreferrer" href={item.url}> <Button variant="primary"><i class="bi bi-box-arrow-up-right me-1"></i>Source</Button></a>}
+           
           </Card.Body>
         </Card>
-        {/* <div className="article-card"> 
-        <a href={item.url} target="_blank" rel="noopener noreferrer">
-          Source Link
-        </a>
-        
-        <br />
-
-        <h5>Title: {item.title}</h5>
-
-        <br />
-        
-         <h5>Topic: {item.topic}</h5>
-
-         <br />
-        
-         <img src={imageUrl} alt="thumbnail" width="250" height="170" />
-         <br />
-
+        {/* 
+        <div className="article-card"> 
+          <a href={item.url} target="_blank" rel="noopener noreferrer"> Source Link </a>
+          <h5>Title: {item.title}</h5>
+          <h5>Topic: {item.topic}</h5>
+          <img src={imageUrl} alt="thumbnail" width="250" height="170" />
           <p>Summary: {item.summary} </p>
-
-        <Link to={`/details/${item.title}`} className="button">Open Details</Link>
-        
-      </div> */}
+          <Link to={`/details/${item.title}`} className="button">Open Details</Link>
+        </div> 
+        */}
       </>
     );
   };
-
-  // console.log(recentDocument)
   return (
 
     <div className="App">
@@ -201,29 +189,31 @@ function EditArticle() {
             Search
           </Button>
         </InputGroup>
-        {/* <label>
-        <input
-          type="radio"
-          value="search"
-          checked={selectedType === 'search'}
-          onChange={() => setSelectedType('search')}
-        />
-        Search
-      </label>
-      <input
-        type="text"
-        placeholder={`Enter ${selectedType === 'search' ? 'context' : 'URL'}`}
-        value={context}
-        onChange={(e) => setContext(e.target.value)}
-      />
-      <button onClick={handleUpdate}>Search</button> */}
         {/* <p>{updateMessage}</p> */}
-        <div className='my-2'>
+        <div className='my-2 results'>
           {!recentDocumentLoading && searchPerformed ? (
             recentDocument ? (
               <>
-
-                {Object.keys(recentDocument).map((topic, index) => {
+                {data
+                  .filter(item => !selectedFilterTopic || item.topic.toLowerCase().includes(selectedFilterTopic.toLowerCase()))
+                  .filter(item => {
+                    const lowerSearch = searchString.toLowerCase()
+                    if (recentDocument[lowerSearch].includes(item.title)) {
+                      for (const EACH_DOCUMENT in recentDocument[lowerSearch]) {
+                        if (recentDocument[lowerSearch][EACH_DOCUMENT] === item.title)
+                          item.rank = EACH_DOCUMENT
+                      }
+                      return item
+                    }
+                    return null
+                  })
+                  // .sort(compareByRank)
+                  .map((item, index) => {
+                    return (
+                      <ArticleCard key={index} item={item} />
+                    )
+                  })}
+                {/* {Object.keys(recentDocument).map((topic, index) => {
                   if (topic.toLowerCase().includes(context.toLowerCase())) {
                     return (
                       <div key={index} >
@@ -259,7 +249,7 @@ function EditArticle() {
                     );
                   }
                   return null;
-                })}
+                })} */}
               </>
 
             ) : (
